# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Frame by Frame - Accounts API",
    "version": "1.0.0",
    "description": "Account sign-up requests, log in, password reset, self-service account management and admin management of users and account requests. Authentication uses an HttpOnly `frame_session` cookie plus an `X-CSRF-Token` header on every state-changing call. 2FA codes are emailed (or printed to the API console in development without SMTP)."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Local"
    }
  ],
...
  "paths": {
...
    "/accounts/signup": {
      "post": {
        "tags": [
          "Sign up & recovery"
        ],
        "summary": "Request a new account",
        "description": "Creates a pending account request that an administrator must approve. The response is identical whether or not the email is already registered.",
        "operationId": "signUpCreateAccountRequest",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SignupRequest"
              },
              "example": {
                "email": "{{signupEmail}}",
                "password": "{{signupPassword}}",
                "displayName": "Postman Tester",
                "message": "Requesting access from Postman."
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Request received",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/password-reset/request": {
      "post": {
        "tags": [
          "Sign up & recovery"
        ],
        "summary": "Email a password reset token",
        "description": "Always returns 202 to avoid account enumeration. In development without SMTP the token is printed to the API console.",
        "operationId": "requestPasswordReset",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetRequest"
              },
              "example": {
                "email": "{{signupEmail}}"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid email",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/password-reset/confirm": {
      "post": {
        "tags": [
          "Sign up & recovery"
        ],
        "summary": "Set a new password with a reset token",
        "description": "Token is single-use and expires after 60 minutes. All sessions are revoked.",
        "operationId": "confirmPasswordReset",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetConfirm"
              },
              "example": {
                "token": "{{resetToken}}",
                "password": "{{signupPassword}}"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Invalid or expired token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/account": {
      "get": {
        "tags": [
          "My account"
        ],
        "summary": "Read own account",
        "operationId": "getMyAccount",
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Account",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "401": {
            "description": "Not signed in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "2FA incomplete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires an authenticated session (2FA complete)."
      },
      "patch": {
        "tags": [
          "My account"
        ],
        "summary": "Update own display name or email",
        "description": "Changing email requires `currentPassword`.\n\nRequires an authenticated session (2FA complete).",
        "operationId": "updateMyAccount",
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequest"
              },
              "example": {
                "displayName": "Renamed in Postman"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Current password required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Email in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "My account"
        ],
        "summary": "Delete own account",
        "description": "The last active administrator cannot be deleted.\n\nRequires an authenticated session (2FA complete).",
        "operationId": "deleteMyAccount",
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeleteAccountRequest"
              },
              "example": {
                "currentPassword": "{{password}}"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "403": {
            "description": "Wrong password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Last administrator",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/account/password": {
      "post": {
        "tags": [
          "My account"
        ],
        "summary": "Change own password",
        "description": "Revokes all sessions; sign in again afterwards.\n\nRequires an authenticated session (2FA complete).",
        "operationId": "changeMyPassword",
        "security": [
          {
            "cookieAuth": [],
            "csrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              },
              "example": {
                "currentPassword": "{{password}}",
                "newPassword": "{{newPassword}}"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Changed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Message"
                }
              }
            }
          },
          "400": {
            "description": "Weak password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Wrong current password",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/account-requests": {
      "get": {
        "tags": [
          "Admin - account requests"
        ],
        "summary": "List account requests",
        "operationId": "listAccountRequests",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "pending",
                "approved",
                "rejected"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 25
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          }
        ],
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountRequestList"
                }
              }
            }
          },
          "403": {
            "description": "Admin only",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Requires an authenticated administrator (2FA complete)."
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "frame_session"
      },
      "csrfToken": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token",
        "description": "Value of `csrfToken` returned by log in / verify."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Message": {
        "type": "object",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
      "SignupRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 256,
            "format": "password"
          },
          "displayName": {
            "type": "string",
            "maxLength": 80
          },
          "message": {
            "type": "string",
            "maxLength": 500,
            "description": "Why access is being requested."
          }
        }
      },
...
      "PasswordResetRequest": {
        "type": "object",
        "required": [
          "email"
        ],
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          }
        }
      },
      "PasswordResetConfirm": {
        "type": "object",
        "required": [
          "token",
          "password"
        ],
        "additionalProperties": false,
        "properties": {
          "token": {
            "type": "string",
            "maxLength": 128
          },
          "password": {
            "type": "string",
            "minLength": 12,
            "maxLength": 256,
            "format": "password"
          }
        }
      },
      "UpdateAccountRequest": {
        "type": "object",
        "minProperties": 1,
        "additionalProperties": false,
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "maxLength": 254
          },
          "displayName": {
            "type": "string",
            "maxLength": 80
          },
          "currentPassword": {
            "type": "string",
            "description": "Required when changing email."
          }
        }
      },
...
      "DeleteAccountRequest": {
        "type": "object",
        "required": [
          "currentPassword"
        ],
        "properties": {
          "currentPassword": {
            "type": "string"
          }
        }
      }
    }
  }
}